typescriptify the serverless-component, adding faq for apex/www subdomain redirection
