+++ .eslintrc.js
@@ -2,7 +2,8 @@ module.exports = {
   root: true,
   extends: ["plugin:prettier/recommended"],
   parserOptions: {
-    ecmaVersion: 2018
+    ecmaVersion: 2018,
+    sourceType: "module"
   },
   overrides: [
     {
@@ -10,8 +11,9 @@ module.exports = {
       files: ["*.ts", "*.tsx"],
       plugins: ["@typescript-eslint"],
       extends: [
-        "plugin:@typescript-eslint/eslint-recommended",
-        "plugin:@typescript-eslint/recommended"
+        "plugin:@typescript-eslint/recommended", // Uses the recommended rules from the @typescript-eslint/eslint-plugin
+        "prettier/@typescript-eslint", // Uses eslint-config-prettier to disable ESLint rules from @typescript-eslint/eslint-plugin that would conflict with prettier
+        "plugin:prettier/recommended" // Enables eslint-plugin-prettier and eslint-config-prettier. This will display prettier errors as ESLint errors. Make sure this is always the last configuration in the extends array.
       ],
       rules: {
         "@typescript-eslint/ban-ts-ignore": "off",
