+++ package.json
@@ -14,10 +14,11 @@
     "api gateway"
   ],
   "scripts": {
+    "prepare": "yarn packages-install",
     "test": "jest --runInBand",
-    "packages-install": "lerna exec -- yarn",
+    "packages-install": "lerna exec yarn",
     "packages-build": "lerna run build",
-    "test:watch": "yarn test -- --watch --collect-coverage=false",
+    "test:watch": "yarn test --watch --collect-coverage=false",
     "publish": "lerna publish --conventional-commits",
     "prerelease": "lerna publish --conventional-commits --conventional-prerelease",
     "graduate": "lerna publish --conventional-commits --conventional-graduate",
@@ -37,14 +38,16 @@
   },
   "homepage": "https://github.com/danielcondemarin/serverless-next.js#readme",
   "devDependencies": {
-    "@babel/preset-typescript": "^7.10.1",
+    "@babel/preset-typescript": "^7.10.4",
     "@sls-next/lambda-at-edge": "link:./packages/lambda-at-edge",
+    "@sls-next/next-aws-cloudfront": "link:./packages/lambda-at-edge-compat",
+    "@types/fs-extra": "^9.0.1",
     "@types/jest": "^26.0.3",
     "@types/react": "^16.9.41",
     "@types/react-dom": "^16.9.8",
     "@types/webpack": "^4.41.18",
-    "@typescript-eslint/eslint-plugin": "^3.4.0",
-    "@typescript-eslint/parser": "^3.4.0",
+    "@typescript-eslint/eslint-plugin": "^3.5.0",
+    "@typescript-eslint/parser": "^3.5.0",
     "adm-zip": "^0.4.16",
     "coveralls": "^3.1.0",
     "eslint": "^7.3.1",
@@ -57,13 +60,12 @@
     "lerna": "^3.22.1",
     "lint-staged": "^10.2.11",
     "next": "^9.4.4",
-    "@sls-next/next-aws-cloudfront": "link:./packages/lambda-at-edge-compat",
     "prettier": "^2.0.5",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
-    "serverless": "^1.74.0",
+    "serverless": "^1.74.1",
     "serverless-offline": "^6.4.0",
-    "typescript": "^3.9.5"
+    "typescript": "^3.9.6"
   },
   "jest": {
     "clearMocks": true,
