+++ packages/lambda-at-edge-compat/next-aws-cloudfront.d.ts
@@ -1,4 +1,3 @@
-// @ts-ignore
 import { CloudFrontResultResponse, CloudFrontRequest } from "aws-lambda";
 import { IncomingMessage, ServerResponse } from "http";
