+++ packages/lambda-at-edge/package.json
@@ -31,13 +31,13 @@
   },
   "homepage": "https://github.com/danielcondemarin/serverless-next.js#readme",
   "devDependencies": {
-    "@types/aws-lambda": "^8.10.56",
+    "@types/aws-lambda": "^8.10.57",
     "@types/execa": "^2.0.0",
     "@types/fs-extra": "^9.0.1",
-    "@types/node": "^14.0.13",
+    "@types/node": "^14.0.14",
     "@types/path-to-regexp": "^1.7.0",
     "ts-loader": "^7.0.5",
-    "typescript": "^3.9.5"
+    "typescript": "^3.9.6"
   },
   "dependencies": {
     "@zeit/node-file-trace": "^0.6.5",
