+++ packages/lambda-at-edge/tests/dynamic-routes-precedence/dynamic-routes-precedence.build.test.ts
@@ -7,7 +7,7 @@ import { cleanupDir } from "../test-utils";
 import {
   OriginRequestDefaultHandlerManifest,
   OriginRequestApiHandlerManifest
-} from "../../src/types";
+} from "../../types";
 
 jest.mock("execa");
