+++ packages/lambda-at-edge/tests/test-utils.ts
@@ -1,7 +1,7 @@
 import path from "path";
 import { remove } from "fs-extra";
 import { CloudFrontOrigin } from "aws-lambda";
-import { OriginRequestEvent } from "../src/types";
+import { OriginRequestEvent } from "../types";
 
 export const cleanupDir = (dir: string): Promise<void> => {
   return remove(dir);
