+++ packages/serverless-component/.d.ts
@@ -1,0 +1,15 @@
+declare module "@serverless/core" {
+  import { Credentials } from "@sls-next/s3-static-assets/src/lib/s3";
+
+  export class Component {
+    load(modulePath: string, moduleName?: string): any;
+    context: {
+      credentials: {
+        aws: Credentials;
+      };
+      instance: {
+        debugMode: boolean;
+      };
+    };
+  }
+}
