+++ packages/serverless-component/__mocks__/@sls-next/cloudfront.js
@@ -1,3 +1,1 @@
-module.exports = {
-  default: jest.fn()
-};
+module.exports = jest.fn();
