+++ packages/serverless-component/__tests__/deploy.test.ts
@@ -1,15 +1,12 @@
-const path = require("path");
-const fse = require("fs-extra");
-const NextjsComponent = require("../serverless");
-const { mockS3 } = require("@serverless/aws-s3");
-const { mockCloudFront } = require("@sls-next/aws-cloudfront");
-const { mockLambda, mockLambdaPublish } = require("@sls-next/aws-lambda");
-const mockCreateInvalidation = require("@sls-next/cloudfront");
-const {
-  DEFAULT_LAMBDA_CODE_DIR,
-  API_LAMBDA_CODE_DIR
-} = require("../constants");
-const { cleanupFixtureDirectory } = require("../lib/test-utils");
+import path from "path";
+import fse from "fs-extra";
+import { mockS3 } from "@serverless/aws-s3";
+import { mockCloudFront } from "@serverless/aws-cloudfront";
+import { mockLambda, mockLambdaPublish } from "@sls-next/aws-lambda";
+import mockCreateInvalidation from "@sls-next/cloudfront";
+import NextjsComponent from "../src/component";
+import { DEFAULT_LAMBDA_CODE_DIR, API_LAMBDA_CODE_DIR } from "../src/constants";
+import { cleanupFixtureDirectory } from "../src/lib/test-utils";
 
 describe("deploy tests", () => {
   let tmpCwd;
@@ -177,7 +174,7 @@ describe("deploy tests", () => {
     });
 
     it("invalidates distribution cache", () => {
-      expect(mockCreateInvalidation.default).toBeCalledWith({
+      expect(mockCreateInvalidation).toBeCalledWith({
         credentials: {
           accessKeyId: "123",
           secretAccessKey: "456"
