--- packages/serverless-component/constants.js
@@ -1,5 +1,0 @@
-module.exports = {
-  BUILD_DIR: ".serverless_nextjs",
-  DEFAULT_LAMBDA_CODE_DIR: ".serverless_nextjs/default-lambda",
-  API_LAMBDA_CODE_DIR: ".serverless_nextjs/api-lambda"
-};
