+++ packages/serverless-component/src/constants.ts
@@ -1,0 +1,3 @@
+export const BUILD_DIR = ".serverless_nextjs";
+export const DEFAULT_LAMBDA_CODE_DIR = ".serverless_nextjs/default-lambda";
+export const API_LAMBDA_CODE_DIR = ".serverless_nextjs/api-lambda";
