+++ packages/serverless-component/types.d.ts
@@ -1,0 +1,40 @@
+import { PublicDirectoryCache } from "@sls-next/s3-static-assets/src/lib/getPublicAssetCacheControl";
+
+export type ServerlessComponentInputs = {
+  build?: BuildOptions;
+  nextConfigDir?: string;
+  useServerlessTraceTarget?: boolean;
+  nextStaticDir?: string;
+  bucketName?: string;
+  bucketRegion?: string;
+  publicDirectoryCache?: PublicDirectoryCache;
+  memory?: number | { defaultLambda?: number; apiLambda?: number };
+  timeout?: number | { defaultLambda?: number; apiLambda?: number };
+  name?: string | { defaultLambda?: string; apiLambda?: string };
+  runtime?: string | { defaultLambda?: string; apiLambda?: string };
+  description?: string;
+  policy?: string;
+  domain?: string | string[];
+  domainType?: "www" | "apex" | "both";
+  cloudfront?: CloudfrontOptions;
+};
+
+type CloudfrontOptions = Record<string, any>;
+
+type BuildOptions = {
+  cwd?: string;
+  enabled?: boolean;
+};
+
+export type LambdaType = "defaultLambda" | "apiLambda";
+
+export type LambdaInput = {
+  description: string;
+  handler: string;
+  code: string;
+  role: Record<string, unknown>;
+  memory: number;
+  timeout: number;
+  runtime: string;
+  name?: string;
+};
